add registry synchronizer tests
